refactor: PostHeader を移動して FC に書き直し
