+++ apps/web/app/(other)/preview/page.tsx
@@ -12,7 +12,7 @@ import Toc from "../posts/[slug]/_components/Toc";
 import TocOnSideContainer from "../posts/[slug]/_components/TocOnSideContainer";
 import { PostEyecatch } from "../posts/[slug]/_components/PostEyecatch";
 import { PostFooterImage } from "../posts/[slug]/_components/PostFooterImage";
-import { PostHeader } from "../posts/[slug]/_components/PostHeader";
+import { PostHeader } from "../posts/[slug]/_post-header";
 
 type Props = {
   searchParams: Promise<ParsedUrlQuery>;
